Rename GameRunner to WormsArmageddon
