+++ CommandLine/src/Commands/Host.cs
@@ -1,6 +1,6 @@
 using System.Threading.Tasks;
 using McMaster.Extensions.CommandLineUtils;
-using Worms.GameRunner;
+using Worms.WormsArmageddon;
 
 namespace Worms.Commands
 {
