+++ CommandLine/src/WormsArmageddon/Windows/WormsRunner.cs
@@ -2,7 +2,7 @@
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace Worms.GameRunner.Windows
+namespace Worms.WormsArmageddon.Windows
 {
     internal class WormsRunner : IWormsRunner
     {
